+++ src/rearmed/generic.js
@@ -3,3 +3,4 @@ require('./generic/isBlank');
 require('./generic/isPresent');
 require('./generic/presence');
 require('./generic/simpleType');
+require('./generic/try');