+++ CHANGELOG.md
@@ -4,7 +4,7 @@
   * Move all Object patches under `Object.prototype.rearmed()` and `Object.rearmed` method to fix issue #2
   * Add `Object#rearmed.add()` and `Object.rearmed.remove()` methods for adding or removing methods to rearmed objects
   * Remove Core section to be replaced with Generic section
-  * Add Generic `isBlank`, `isPresent`, `presence`, `simpleType`
+  * Add Generic `isBlank`, `isPresent`, `presence`, `simpleType`, `try`
   * Add Array `range` and `tap` methods
   * Fix `Object#all` when called with no callback function. Now defaults to if all values are truthy.
   * Remove `Array#notEmpty`, `Object#notEmpty`, and `String#notEmpty`